update Float in ScreenPropertyHelper to net6

+++ Yang.Maui.Helper/Device/Screen/UnitHelper.cs
@@ -25,7 +25,7 @@ namespace Yang.Maui.Helper.Device.Screen
                 if (density == 0)
                 {
 #if __IOS__ || ANDROID || WINDOWS
-                    density = ScreenPropertyHelper.GetScreenDensity();
+                    density = (float)ScreenPropertyHelper.GetScreenDensity();
 #endif
                 }
                 return density;
